Adding textmode wrapper in C
The ammount of time I have spent making cargo pass proper flags to
rustc is just insane. In the end I DID EVEN SUCCEEDED! This build
system is so broken I have to words for this. This is the ONLY and very
reason I decided to go with C (atleast for now)